Add test to compare against randomkit Zipf
Note that this test does not currently fail if the distributions are
identified as different. The reason for this (as indicated by the
comments) is that it is unclear that the alpha for this implementation,
and the alpha as used by ranomkit's Zipf are actually the same
parameter. Thus, even if they are given the same value, it is unclear if
the resulting distributions should be the same.

Visual inspection suggests that both generate a Zipf distribution, but
that they generate *different* distribution for alpha=1.07.

Suggestions on how to fix this + other ways to test are appreciated.